pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.12",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "html5lib>=1.1",
    "imblearn>=0.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.23",
    "langchain-text-splitters>=0.3.8",
    "lxml>=5.4.0",
    "matplotlib>=3.10.3",
    "nbformat>=5.10.4",
    "numpy>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pydantic>=2.11.6",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
    "statsmodels>=0.14.4",
    "streamlit>=1.45.1",
    "wikipedia>=1.4.0",
]
